refactor: domain repository QueyrDSL화
