+++ src/main/java/com/umc/linkyou/repository/classification/domainRepository/DomainRepository.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.repository.classification;
+package com.umc.linkyou.repository.classification.domainRepository;
 
 import com.umc.linkyou.domain.classification.Domain;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,10 +10,5 @@ import java.util.Optional;
 
 @Repository
 public interface DomainRepository extends JpaRepository<Domain, Long> {
-    Optional<Domain> findByDomainTail(String domainTail);
 
-    Optional<Domain> findById(long l);
-
-    // N+1 제거용 일괄 조회
-    List<Domain> findByDomainTailIn(Collection<String> domainTails);
 }
