+++ src/main/java/com/umc/linkyou/service/DomainServiceImpl.java
@@ -3,10 +3,9 @@ package com.umc.linkyou.service;
 import com.umc.linkyou.apiPayload.code.status.ErrorStatus;
 import com.umc.linkyou.apiPayload.exception.GeneralException;
 import com.umc.linkyou.awsS3.AwsS3Service;
-import com.umc.linkyou.converter.AwsS3Converter;
 import com.umc.linkyou.converter.DomainConverter;
 import com.umc.linkyou.domain.classification.Domain;
-import com.umc.linkyou.repository.classification.DomainRepository;
+import com.umc.linkyou.repository.classification.domainRepository.DomainRepository;
 import com.umc.linkyou.web.dto.DomainDTO;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
